# src/main/resources/static/PlusMin-api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:3045/api/v1
  description: Generated server url
security:
- bearerAuth: []
paths:
  /periodes/hulpvrager/{hulpvragerId}/wijzig-periode-opening/{periodeId}:
    put:
      tags:
      - periode-controller
      summary: PUT wijzig een periode voor hulpvrager
      operationId: wijzigPeriodeOpeningVoorHulpvrager
      parameters:
      - name: hulpvragerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: periodeId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/SaldoDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /periodes/hulpvrager/{hulpvragerId}/sluiten/{periodeId}:
    put:
      tags:
      - periode-controller
      summary: PUT sluit een periode voor hulpvrager
      operationId: sluitPeriodeVoorHulpvrager
      parameters:
      - name: hulpvragerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: periodeId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/SaldoDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /periodes/hulpvrager/{hulpvragerId}/opruimen/{periodeId}:
    put:
      tags:
      - periode-controller
      summary: PUT ruim een periode op (verwijder de betalingen) voor hulpvrager
      operationId: ruimPeriodesOpVoorHulpvrager
      parameters:
      - name: hulpvragerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: periodeId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /periodes/hulpvrager/{hulpvragerId}/heropenen/{periodeId}:
    put:
      tags:
      - periode-controller
      summary: PUT heropen een periode (en verwijder de saldi) voor hulpvrager
      operationId: heropenPeriodeVoorHulpvrager
      parameters:
      - name: hulpvragerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: periodeId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /rekeningen/hulpvrager/{hulpvragerId}:
    post:
      tags:
      - rekening-controller
      operationId: creeerNieuweRekeningVoorHulpvrager
      parameters:
      - name: hulpvragerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/RekeningGroepDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /stand/hulpvrager/{hulpvragerId}/periode/{periodeId}/openingsbalans:
    get:
      tags:
      - stand-controller
      summary: GET de stand voor hulpvrager op datum
      operationId: getOpeningsBalansVoorPeriode
      parameters:
      - name: hulpvragerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: periodeId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SaldoDTO"
  /stand/hulpvrager/{hulpvragerId}/datum/{datum}:
    get:
      tags:
      - stand-controller
      summary: GET de stand voor hulpvrager op datum
      operationId: getStandOpDatumVoorHulpvrager
      parameters:
      - name: hulpvragerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: datum
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/StandDTO"
  /stand/hulpvrager/{hulpvragerId}/checkspaarsaldi:
    get:
      tags:
      - stand-controller
      summary: GET de spaarsaldi controle voor hulpvrager
      operationId: checkSaldi
      parameters:
      - name: hulpvragerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /rekeningen/hulpvrager/{hulpvragerId}/periode/{periodeId}:
    get:
      tags:
      - rekening-controller
      summary: GET de geldige rekeningen in een periode
      operationId: getAlleRekeningenPerBetalingsSoortVoorHulpvrager
      parameters:
      - name: hulpvragerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: periodeId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /rekeningen/hulpvrager/{hulpvragerId}/periode/{periodeId}/cashflow:
    get:
      tags:
      - rekening-controller
      summary: GET de cashflow in een periode
      operationId: getCashflowVoorHulpvrager
      parameters:
      - name: hulpvragerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: periodeId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /periodes/hulpvrager/{hulpvragerId}:
    get:
      tags:
      - periode-controller
      summary: GET de periodes voor hulpvrager
      operationId: getPeriodesVoorHulpvrager
      parameters:
      - name: hulpvragerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /periodes/hulpvrager/{hulpvragerId}/sluitvoorstel/{periodeId}:
    get:
      tags:
      - periode-controller
      summary: GET voorstel voor het sluiten van een periode voor hulpvrager
      operationId: sluitPeriodeVoorstelVoorHulpvrager
      parameters:
      - name: hulpvragerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: periodeId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
components:
  schemas:
    AflossingDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        startDatum:
          type: string
        eindBedrag:
          type: string
        dossierNummer:
          type: string
        notities:
          type: string
      required:
      - dossierNummer
      - eindBedrag
      - id
      - notities
      - startDatum
    SaldoDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        rekeningGroepNaam:
          type: string
        rekeningGroepSoort:
          type: string
          enum:
          - BETAALREKENING
          - SPAARREKENING
          - CONTANT
          - CREDITCARD
          - INKOMSTEN
          - UITGAVEN
          - AFLOSSING
          - SPAARPOT
          - RESERVERING_BUFFER
        budgetType:
          type: string
          enum:
          - INKOMSTEN
          - VAST
          - CONTINU
          - SPAREN
        rekeningNaam:
          type: string
        aflossing:
          $ref: "#/components/schemas/AflossingDTO"
        spaartegoed:
          $ref: "#/components/schemas/SpaartegoedDTO"
        sortOrder:
          type: integer
          format: int32
        openingsBalansSaldo:
          type: number
        openingsReserveSaldo:
          type: number
        openingsOpgenomenSaldo:
          type: number
        achterstand:
          type: number
        budgetMaandBedrag:
          type: number
        budgetBetaalDag:
          type: integer
          format: int32
        budgetAanvulling:
          type: string
          enum:
          - TOT
          - MET
          - IN
          - UIT
        betaling:
          type: number
        reservering:
          type: number
        opgenomenSaldo:
          type: number
        correctieBoeking:
          type: number
        achterstandOpPeilDatum:
          type: number
        budgetPeilDatum:
          type: string
        budgetOpPeilDatum:
          type: number
        betaaldBinnenBudget:
          type: number
        minderDanBudget:
          type: number
        meerDanBudget:
          type: number
        meerDanMaandBudget:
          type: number
        restMaandBudget:
          type: number
      required:
      - achterstand
      - betaling
      - budgetMaandBedrag
      - correctieBoeking
      - id
      - openingsBalansSaldo
      - openingsOpgenomenSaldo
      - openingsReserveSaldo
      - opgenomenSaldo
      - rekeningGroepNaam
      - rekeningNaam
      - reservering
      - sortOrder
    SpaartegoedDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        doelDatum:
          type: string
        doelBedrag:
          type: string
        notities:
          type: string
      required:
      - id
      - notities
...
    Periode:
      type: object
      properties:
        id:
          type: integer
          format: int64
        periodeStartDatum:
          type: string
          format: date
        periodeEindDatum:
          type: string
          format: date
        periodeStatus:
          type: string
          enum:
          - HUIDIG
          - OPEN
          - GESLOTEN
          - OPGERUIMD
      required:
      - id
      - periodeEindDatum
      - periodeStartDatum
      - periodeStatus
    RekeningDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        naam:
          type: string
        rekeningGroepNaam:
          type: string
        sortOrder:
          type: integer
          format: int32
        bankNaam:
          type: string
        vanPeriode:
          $ref: "#/components/schemas/Periode"
        totEnMetPeriode:
          $ref: "#/components/schemas/Periode"
        gekoppeldeRekening:
          type: string
        budgetPeriodiciteit:
          type: string
        saldo:
          type: number
        reserve:
          type: number
        budgetBedrag:
          type: number
        budgetVariabiliteit:
          type: integer
          format: int32
        maanden:
          type: array
          items:
            type: integer
            format: int32
          uniqueItems: true
        budgetBetaalDag:
          type: integer
          format: int32
        budgetAanvulling:
          type: string
          enum:
          - TOT
          - MET
          - IN
          - UIT
        budgetMaandBedrag:
          type: number
        aflossing:
          $ref: "#/components/schemas/AflossingDTO"
        spaartegoed:
          $ref: "#/components/schemas/SpaartegoedDTO"
      required:
      - id
      - naam
    RekeningGroepDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        naam:
          type: string
        rekeningGroepSoort:
          type: string
        rekeningGroepIcoonNaam:
          type: string
        sortOrder:
          type: integer
          format: int32
        budgetType:
          type: string
        rekeningen:
          type: array
          items:
            $ref: "#/components/schemas/RekeningDTO"
      required:
      - id
      - naam
      - rekeningGroepSoort
      - rekeningen
      - sortOrder
...
    ResultaatSamenvattingOpDatumDTO:
      type: object
      properties:
        percentagePeriodeVoorbij:
          type: integer
          format: int64
        openingsReservePotjesVoorNuSaldo:
          type: number
        budgetMaandInkomstenBedrag:
          type: number
        besteedTotPeilDatum:
          type: number
        gespaardTotPeilDatum:
          type: number
        nogNodigNaPeilDatum:
          type: number
        actueleBuffer:
          type: number
        extraGespaardTotPeilDatum:
          type: number
      required:
      - actueleBuffer
      - besteedTotPeilDatum
      - budgetMaandInkomstenBedrag
      - extraGespaardTotPeilDatum
      - gespaardTotPeilDatum
      - nogNodigNaPeilDatum
      - openingsReservePotjesVoorNuSaldo
      - percentagePeriodeVoorbij
    StandDTO:
      type: object
      properties:
        periodeStartDatum:
          type: string
          format: date
        peilDatum:
          type: string
          format: date
        datumLaatsteBetaling:
          type: string
          format: date
        budgetHorizon:
          type: string
          format: date
        reserveringsHorizon:
          type: string
          format: date
        resultaatOpDatum:
          type: array
          items:
            $ref: "#/components/schemas/SaldoDTO"
        resultaatSamenvattingOpDatum:
          $ref: "#/components/schemas/ResultaatSamenvattingOpDatumDTO"
        geaggregeerdResultaatOpDatum:
          type: array
          items:
            $ref: "#/components/schemas/SaldoDTO"
      required:
      - budgetHorizon
      - geaggregeerdResultaatOpDatum
      - peilDatum
      - periodeStartDatum
      - reserveringsHorizon
      - resultaatOpDatum
      - resultaatSamenvattingOpDatum
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT